handleQuickSearch passed as a prop, logic moved to ListContainer
