+++ src/NativeBase/Components/ProjectTagButtonsFilter.tsx
@@ -1,25 +1,15 @@
-import { navigate, RootStackParamList } from '@/Navigators/utils'
 import {
   Projects,
   ProjectSector,
   ProjectsSearchField,
   ProjectTechnology,
 } from '@/Services/modules/projects'
-import {
-  RoleGroupName,
-  roleGroups,
-} from '@/Services/modules/projects/roleGroups'
+import { RoleGroupName } from '@/Services/modules/projects/roleGroups'
 import { ProjectsState } from '@/Store/Projects'
-import { fuzzySearchByArray, searchByArray } from '@/Utils/Search'
-import Fuse from 'fuse.js'
 import { VStack } from 'native-base'
-import React, { useState } from 'react'
+import React, { FC, useState } from 'react'
 import { useSelector } from 'react-redux'
-import {
-  ListRouteParams,
-  ListSearch,
-  ListType,
-} from '../Containers/ListContainer'
+import { ListSearch } from '../Containers/ListContainer'
 import ChoicesList, {
   ChoicesListChoice,
   ChoicesListColour,
@@ -30,8 +20,16 @@ import TagButtons from './TagButtons'
 export interface ProjectSearch extends ListSearch {
   results: Projects // the projects results for this search
 }
+export interface ProjectsTagButtonsFilterProps {
+  handleQuickSearchSubmit: (
+    searchField: ProjectsSearchField,
+    searchQueryChoice: string,
+  ) => void
+}
 
-const ProjectsTagButtonsFilter = () => {
+const ProjectsTagButtonsFilter: FC<ProjectsTagButtonsFilterProps> = ({
+  handleQuickSearchSubmit,
+}) => {
   // Fetch all projects from the store
   const allProjects = useSelector(
     (state: { projects: ProjectsState }) => state.projects.projects,
@@ -84,77 +82,6 @@ const ProjectsTagButtonsFilter = () => {
   // Define colour and style to use for quick search options
   const quickSearchListColour = ChoicesListColour.primary
   const quickSearchListStyle = ChoicesListFontStyle.mediumLight
-
-  // Ensure job title searches find related roles
-  const getRelatedRoles = (
-    possibleRoleSearchQuery: string,
-  ): string[] | undefined => {
-    const fuse = new Fuse(roleGroups, {
-      keys: ['roleNames'],
-      minMatchCharLength: 2,
-      threshold: 0.1,
-    })
-
-    const fuseResults = fuse.search(possibleRoleSearchQuery)
-
-    if (fuseResults.length) {
-      const roles = []
-
-      for (const fuseResult of fuseResults) {
-        for (const role of fuseResult.item.roleNames) {
-          roles.push(role)
-        }
-      }
-
-      return roles
-    }
-
-    return undefined
-  }
-
-  const handleQuickSearchSubmit = (
-    searchField: ProjectsSearchField,
-    searchQueryChoice: string,
-  ) => {
-    let searchQueries = [] as string[]
-    let results = [] as Projects
-
-    searchQueries.push(searchQueryChoice)
-
-    if (searchField === 'role') {
-      const relatedRoles = getRelatedRoles(searchQueryChoice)
-
-      if (relatedRoles?.length) {
-        searchQueries = searchQueries.concat(relatedRoles)
-      }
-      results = fuzzySearchByArray(
-        allProjects,
-        [searchField],
-        searchQueries,
-      ) as Projects // we need to use fuzzy search as the roles names are not exact (charities use different ways of naming roles)
-    } else {
-      results = searchByArray(
-        allProjects,
-        searchField,
-        searchQueries,
-      ) as Projects // here we do not want to use fuzzy search as it would include unwanted results
-    }
-
-    const description = `${
-      searchField === 'sector' ? 'cause' : searchField
-    } "${searchQueryChoice}"`
-
-    navigate(
-      'Projects' as keyof RootStackParamList,
-      {
-        type: ListType.Projects,
-        search: {
-          results,
-          description,
-        } as ProjectSearch,
-      } as ListRouteParams,
-    )
-  }
 
   /**
    * Handle tag press logic
